requirements.txt
Django>=3.2,<5.0
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
